+++ packages/datocms-plugin-table/src/components/HeadingEditor/style.module.css
@@ -1,7 +1,3 @@
-.textEditor {
-  min-width: 100%;
-}
-
 .editor {
   position: relative;
 }
